fix(ci): use lockfile for building & ignore local node_modules
the current dockerfile was using the `npm i` command instead of
`npm ci`, which allowed it to automatically & silently update
dependencies.
this allowed for issues, where it could update a dependency to one that
would break the build (but only if it was a minor/patch update, i.e. the
dependency devs made a mistake)
coincidentally, `adonis-autoswagger` devs appear to have just made that
exact mistake, adding a new required field on a public type in a minor
update.
this commit fixes this issue by using `npm ci` instead.

addtionally, there was no `.dockerignore` file, which combined with the
`ADD . .` command just before the build in the `Dockerfile` caused the
build to actually use the dependencies installed in the local repo,
outside of the container image build environment, by copying the local
`node_modules` directory into the build env.
this made the bug explained earlier not reproducible locally, if your
local repo had dependencies already installed.
reproducing the bug required the cleanup of the repo by running
`git reset --hard` and `git clean -fdx`.
this commit fixes this issue by adding a `.dockerignore` which mostly
mirrors the directories specified `.gitignore`, including the
`node_modules` directory.